Rollback #2

+++ src/yadi/dlms/DlmsType.java
@@ -44,7 +44,7 @@ public enum DlmsType {
 	UINT32(6,4),
 	UINT64(21,8);
 	
-	public final int tag;
+	public final byte tag;
 	final int size;
 	
 	DlmsType(int tag, int size) {
@@ -52,7 +52,7 @@ public enum DlmsType {
 		this.size = size;
 	}
 	
-	public static DlmsType fromTag(int tag) throws DlmsException {
+	public static DlmsType fromTag(byte tag) throws DlmsException {
 		for (DlmsType type: values()) {
 			if (type.tag == tag) {
 				return type;
